Replaced old templates and added constants.py
Changed some templates, made constants.py to store constants and improve the readability of the code, added more features and capabilities, such as an account system with password hashing and saving bonuses (bonuses can't be used to pay yet). Fixed and changed many functions, tried to make the code clean in general.